made sys argv optional

+++ server.py
@@ -16,7 +16,7 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import func
 from sys import argv
 
-script, mode = argv
+
 
 app = Flask(__name__)
 
@@ -586,10 +586,13 @@ def connect_to_db(app, db, db_name):
     db.app = app
     db.init_app(app)
 
+
     with app.app_context(): # http://stackoverflow.com/questions/19437883/when-scattering-flask-models-runtimeerror-application-not-registered-on-db-w
         # if in reseed mode, delete database so can be reseeded.
-        if mode == "reseed":
-            os.remove(db_name)
+        if len(argv) > 1:
+            script, mode = argv
+            if mode == "reseed":
+                os.remove(db_name)
         # if database doesn't exist yet, creates database
         if os.path.isfile(db_name) != True:
             db.create_all()
